logistics-route-planner/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
groq
faiss-cpu
numpy
SQLAlchemy
psycopg2-binary
alembic
sentence-transformers
langchain
langchain-community
langchain-openai
langchain-google-genai
langgraph
duckduckgo-search
wikipedia
requests
beautifulsoup4
lxml
tavily-python
arxiv
pymupdf
